wip
next, for each role (lead and follow) we need to align the ankles at frame 0, but then we need to find the discontinuity, and then rotate the entire animation around the frame 0 ankle floor point (so a revolution around Y, centered at this point), such that the ankle ground at the discontinuity point is as close as possible to the 3D ankle of the pose at that position

for example, if frame0 lead right ankle is aligned at point 1,0,2, and there is a discontinuity at frame 2000, we should get the ground ankle at frame 2000. if the ground ankle (lets say left) is at position -5, 0, 17, and the 3D left ankle is at xyz, the entire animation up to frame 2000 needs to be rotated around Y centered at the frame 0 ground ankle so that the frame 2000 ankle is as close as possible to the ground ankle position. this will cure the world orientation for both poses